+++ src/gui-qt4/libs/seiscomp3/gui/core/scheme.cpp
@@ -449,6 +449,7 @@ Scheme::Precision::Precision() {
 	depth = 0;
 	distance = 1;
 	location = 2;
+	magnitude = 2;
 	pickTime = 1;
 	traceValues = 1;
 	rms = 1;
@@ -631,6 +632,7 @@ void Scheme::fetch() {
 	READ_INT(precision.depth);
 	READ_INT(precision.distance);
 	READ_INT(precision.location);
+	READ_INT(precision.magnitude);
 	READ_INT(precision.pickTime);
 	READ_INT(precision.traceValues);
 	READ_INT(precision.rms);